refactor: move some line insertion code into a shared BlockPatcher
Fixes #196

This should be useful for a later commit that works with loops in the normalize
stage.
